refactor: ComponentEditor 타입 Img에서 badge로 변경
