+++ src/types/index.ts
@@ -1,3 +1,3 @@
-export * from './imgComponentType';
+export * from './badgeComponentType';
 export * from './textComponentType';
 export * from './component';
